requirements.txt
torch
torchvision

pillow
transformers
supervision
ultralytics
clip
opencv-python
scikit-image
streamlit-drawable-canvas

timm
onnxruntime